+++ arduino_scripts/Motor_Shield_Controller/src/SerialInts.h
@@ -6,13 +6,13 @@
 
 class SerialInts {
   public:
-    SerialInts(char sep, int size, unsigned long wait);
+    SerialInts(char sep, long size, unsigned long wait);
     ~SerialInts();
     void start_serial();
     void scan();
     void print_values();
-    int length();
-    int getInt();
+    long length();
+    long getInt();
     void dump_values(char sep);
     void print_sum();
     void print_mean();
@@ -24,8 +24,8 @@ class SerialInts {
     bool _got_byte;
     char _char_code;
     unsigned long _wait;
-    QueueList<int> _values;
-    int _curr_byte,_curr_int;
+    QueueList<long> _values;
+    long _curr_byte,_curr_int;
     char _sep,_in_byte;
     void parse_byte();
     void demo();
